Update readme.md
Request to include the styles for the ion-datetime modal in Angular in the document code to reflect the demo example. Using the code straight from the example makes you think that ion-datetime is broken. Only after digging through the examples it was found out that you need to provide your own custom CSS. I assure this will save many new users time and frustration by including this.